Add Hover Transparency
My players constantly voiced complaints about the notification covering their long-winded speeches and character interactions.
I already applied this to my game but making a PR here so if desired, it's applied for everyone. Otherwise, please see this as a suggestion to fix the issue of having 1/4 of the chat window go away anytime someone else types.